+++ RadioControl/app/src/main/java/com/nikhilparanjape/radiocontrol/fragments/AboutFragment.java
@@ -22,6 +22,18 @@ import com.nikhilparanjape.radiocontrol.R;
 import com.nikhilparanjape.radiocontrol.activities.TutorialActivity;
 import com.nikhilparanjape.radiocontrol.rootUtils.Utilities;
 
+import org.w3c.dom.Document;
+import org.xml.sax.SAXException;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.MalformedURLException;
+import java.net.URL;
+
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+
 import it.gmariotti.changelibs.library.view.ChangeLogRecyclerView;
 
 /**
@@ -39,13 +51,11 @@ public class AboutFragment extends PreferenceFragment {
         Utilities util = new Utilities();
 
 
-
         final Context c = getActivity();
 
-        if(Utilities.isConnected(c)){
+        if (Utilities.isConnected(c)) {
             getPreferenceScreen().findPreference("source").setEnabled(true);
-        }
-        else{
+        } else {
             getPreferenceScreen().findPreference("source").setEnabled(false);
         }
 
@@ -54,26 +64,25 @@ public class AboutFragment extends PreferenceFragment {
         versionPref.setSummary("v" + cs);
         versionPref.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
             int z = 0;
+
             public boolean onPreferenceClick(Preference preference) {
                 SharedPreferences sp = c.getSharedPreferences(PRIVATE_PREF, Context.MODE_PRIVATE); //Initializes prefs.xml
                 SharedPreferences.Editor editor = sp.edit();//Initializes xml editor
                 z++;
-                Log.d("RadioControl",(7-z)+ " steps away from easter egg");
-                if(z >= 7){
-                    if (!sp.getBoolean("isDeveloper",false)) {
+                Log.d("RadioControl", (7 - z) + " steps away from easter egg");
+                if (z >= 7) {
+                    if (!sp.getBoolean("isDeveloper", false)) {
                         Toast.makeText(getActivity(), R.string.dev_activated, Toast.LENGTH_LONG).show();
-                        z=0;
-                        Log.d("RadioControl","Developer features activated");
+                        z = 0;
+                        Log.d("RadioControl", "Developer features activated");
 
 
                         editor.putBoolean("isDeveloper", true); //Puts the boolean into prefs.xml
                         editor.apply(); //Ends writing to prefs file
-                    }
-                    else if(sp.getBoolean("isDeveloper",false)){
+                    } else if (sp.getBoolean("isDeveloper", false)) {
                         Toast.makeText(getActivity(), R.string.dev_deactivated, Toast.LENGTH_LONG).show();
-                        z=0;
-                        Log.d("RadioControl",c.getString(R.string.dev_deactivated));
-
+                        z = 0;
+                        Log.d("RadioControl", c.getString(R.string.dev_deactivated));
 
 
                         editor.putBoolean("isDeveloper", false); //Puts the boolean into prefs.xml
@@ -83,6 +92,7 @@ public class AboutFragment extends PreferenceFragment {
                 }
                 return false;
             }
+
         });
 
         Preference myPref = findPreference("changelog");
@@ -110,7 +120,23 @@ public class AboutFragment extends PreferenceFragment {
         });
 
 
-
+    }
+    public static void getUpdate(){
+        Document doc;
+        try{
+            URL xmlURL = new URL("http://nikhilp.org/radiocontrol/backend/update_check.xml");
+            InputStream xml = xmlURL.openStream();
+            DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+            DocumentBuilder db = dbf.newDocumentBuilder();
+            doc = db.parse(xml);
+            xml.close();
+        } catch (SAXException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } catch (ParserConfigurationException e) {
+            e.printStackTrace();
+        }
 
 
     }
@@ -119,7 +145,7 @@ public class AboutFragment extends PreferenceFragment {
             LayoutInflater inflater = LayoutInflater.from(c);//Creates layout inflator for dialog
             WebView view = (WebView) inflater.inflate(R.layout.dialog_licenses, null);//Initializes the view for whats new dialog
             AlertDialog.Builder builder = new AlertDialog.Builder(c);//creates alertdialog
-            view.loadUrl("https://nikhilp.org/radiocontrol/opensource/index.html");
+            view.loadUrl("http://nikhilp.org/radiocontrol/opensource/index.html");
             builder.setView(view).setTitle(R.string.open_source_title)//sets title
                     .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
                         @Override
@@ -127,6 +153,7 @@ public class AboutFragment extends PreferenceFragment {
                             dialog.dismiss();
                         }
                     });
+
             builder.create().show();
         } else{
             Snackbar.make(getView(), "No internet connection found", Snackbar.LENGTH_LONG)